# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/assets:
    get:
      tags:
        - Assets
      summary: List Assets
      description: Get a paginated list of Assets with optional filtering and sorting
      operationId: listAssets
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Assets
      summary: Create Asset
      description: Create a new Asset
      operationId: createAsset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAssetRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Asset'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    Asset:
      type: object
      required:
        - id
        - company_id
        - asset_category_id
        - asset_name
        - asset_code
        - gross_purchase_amount
        - salvage_value
        - useful_life_months
        - opening_accumulated_depreciation
        - purchase_date
        - accumulated_depreciation
        - net_book_value
        - status
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        asset_category_id:
          type: string
          format: uuid
        asset_name:
          type: string
          maxLength: 140
        asset_code:
          type: string
          maxLength: 40
        item_id:
          nullable: true
          type: string
          format: uuid
        branch_id:
          nullable: true
          type: string
          format: uuid
        gross_purchase_amount:
          type: string
        salvage_value:
          type: string
        useful_life_months:
          type: integer
          format: int32
        opening_accumulated_depreciation:
          type: string
        purchase_date:
          type: string
          format: date-time
        available_for_use_date:
          nullable: true
          type: string
          format: date-time
        accumulated_depreciation:
          type: string
        net_book_value:
          type: string
        status:
          type: string
        metadata:
          type: object
...
    CreateAssetRequest:
      type: object
      required:
        - company_id
        - asset_category_id
        - asset_name
        - asset_code
        - gross_purchase_amount
        - salvage_value
        - useful_life_months
        - opening_accumulated_depreciation
        - purchase_date
        - accumulated_depreciation
        - net_book_value
        - status
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        asset_category_id:
          type: string
          format: uuid
        asset_name:
          type: string
          maxLength: 140
        asset_code:
          type: string
          maxLength: 40
        item_id:
          nullable: true
          type: string
          format: uuid
        branch_id:
          nullable: true
          type: string
          format: uuid
        gross_purchase_amount:
          type: string
        salvage_value:
          type: string
        useful_life_months:
          type: integer
          format: int32
        opening_accumulated_depreciation:
          type: string
        purchase_date:
          type: string
          format: date-time
        available_for_use_date:
          nullable: true
          type: string
          format: date-time
        accumulated_depreciation:
          type: string
        net_book_value:
          type: string
        status:
          type: string
        metadata:
          type: object
...
    AssetList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Asset'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []